frontend: redux/filterSlice: Refactor filter reducers/actions into slice
Moved related filter functionality from lib/util into there too.

There's also some tests in filterSlice.test.ts now.
