+++ frontend/src/components/App/Notifications/List.stories.tsx
@@ -3,7 +3,7 @@ import { Meta, Story } from '@storybook/react/types-6-0';
 import helpers from '../../../helpers';
 import { Notification } from '../../../lib/notification';
 import { initialState as CONFIG_INITIAL_STATE } from '../../../redux/configSlice';
-import { INITIAL_STATE as FILTER_INITIAL_STATE } from '../../../redux/reducers/filter';
+import { initialState as FILTER_INITIAL_STATE } from '../../../redux/filterSlice';
 import { INITIAL_STATE as UI_INITIAL_STATE } from '../../../redux/reducers/ui';
 import { TestContext } from '../../../test';
 import NotificationList from './List';
